+++ examples/audio.js
@@ -9,6 +9,8 @@ loadSound("OtherworldlyFoe", "sounds/OtherworldlyFoe.mp3");
 // the music might not autoplay cuz some browser won't allow audio start before any user interaction
 const music = play("OtherworldlyFoe", { loop: true, });
 
+addSky();
+
 // adjust global volume
 volume(0.5);