+++ examples/button.js
@@ -16,9 +16,9 @@ function addButton(txt, p, f) {
 		if (btn.isHovered()) {
 			const t = time() * 10;
 			btn.color = rgb(
-				wave(0, 1, t),
-				wave(0, 1, t + 2),
-				wave(0, 1, t + 4),
+				wave(0, 255, t),
+				wave(0, 255, t + 2),
+				wave(0, 255, t + 4),
 			);
 			btn.scale = vec2(1.2);
 			cursor("pointer");
@@ -28,7 +28,7 @@ function addButton(txt, p, f) {
 		} else {
 			cursor("default");
 			btn.scale = vec2(1);
-			btn.color = rgb(1, 1, 1);
+			btn.color = rgb();
 		}
 	});