+++ examples/kaboom.js
@@ -1,2 +1,3 @@
 kaboom();
+keyPress("space", () => addKaboom(mousePos()));
 mouseClick(() => addKaboom(mousePos()));